fix: minor improves
