+++ optimum_transformers/benchmark/__init__.py
@@ -1,4 +1,4 @@
-
+# Benchmark
 from .base import (
     Benchmark
 )
