+++ optimum_transformers/pipelines/__init__.py
@@ -2,64 +2,44 @@
 # There's no way to ignore "F401 '...' imported but unused" warnings in this
 # module, but to preserve other warnings. So, don't check this module at all.
 
-import io
-import json
 import os
 
-# coding=utf-8
-# Copyright 2018 The HuggingFace Inc. team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 import warnings
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Union
-from onnxruntime import GraphOptimizationLevel, InferenceSession, SessionOptions, get_all_providers
+from typing import Any, Dict, Optional, Union
+from onnxruntime import GraphOptimizationLevel, InferenceSession, SessionOptions
 
 from transformers.configuration_utils import PretrainedConfig
 from transformers.feature_extraction_utils import PreTrainedFeatureExtractor
-from transformers.file_utils import http_get, is_tf_available, is_torch_available
+from transformers.file_utils import is_tf_available, is_torch_available
 from transformers.models.auto.configuration_auto import AutoConfig
 from transformers.models.auto.feature_extraction_auto import FEATURE_EXTRACTOR_MAPPING, AutoFeatureExtractor
 from transformers.models.auto.tokenization_auto import TOKENIZER_MAPPING, AutoTokenizer
 from transformers.tokenization_utils import PreTrainedTokenizer
 from transformers.utils import logging
+from transformers.pipelines import get_task, check_task
 
 from transformers import pipeline as default_pipeline
 
 from optimum.onnxruntime import ORTConfig, ORTQuantizer
 
-# from .audio_classification import AudioClassificationPipeline
-# from .automatic_speech_recognition import AutomaticSpeechRecognitionPipeline
 from transformers.pipelines.base import (
-    ArgumentHandler,
-    CsvPipelineDataFormat,
-    JsonPipelineDataFormat,
-    PipedPipelineDataFormat,
     Pipeline,
-    PipelineDataFormat,
-    PipelineException,
     get_default_model,
     infer_framework_load_model,
 )
 
-# from .conversational import Conversation, ConversationalPipeline
+from .base import (
+    _warmup_onnx_graph,
+    _forward_onnx,
+    create_model_for_providers,
+    _create_quantized_graph,
+    _export_onnx_graph,
+)
+
 from .feature_extraction import OptimumFeatureExtractionPipeline
 from .fill_mask import OptimumFillMaskPipeline
-# from .image_classification import ImageClassificationPipeline
-# from .image_segmentation import ImageSegmentationPipeline
-# from .object_detection import ObjectDetectionPipeline
+
 from .question_answering import OptimumQuestionAnsweringPipeline
-# from .table_question_answering import TableQuestionAnsweringArgumentHandler, TableQuestionAnsweringPipeline
-# from .text2text_generation import SummarizationPipeline, Text2TextGenerationPipeline, TranslationPipeline
 from .text_classification import OptimumTextClassificationPipeline
 from .text_generation import OptimumTextGenerationPipeline
 from .token_classification import (
@@ -67,25 +47,15 @@ from .token_classification import (
 )
 from .zero_shot_classification import OptimumZeroShotClassificationPipeline
 
-# from .zero_shot_image_classification import ZeroShotImageClassificationPipeline
-
-
 if is_tf_available():
     import tensorflow as tf
 
     from transformers.models.auto.modeling_tf_auto import (
-        TF_MODEL_FOR_QUESTION_ANSWERING_MAPPING,
-        TF_MODEL_FOR_SEQ_TO_SEQ_CAUSAL_LM_MAPPING,
-        TF_MODEL_FOR_SEQUENCE_CLASSIFICATION_MAPPING,
-        TF_MODEL_FOR_TOKEN_CLASSIFICATION_MAPPING,
-        TF_MODEL_WITH_LM_HEAD_MAPPING,
         TFAutoModel,
         TFAutoModelForCausalLM,
         TFAutoModelForMaskedLM,
         TFAutoModelForQuestionAnswering,
-        TFAutoModelForSeq2SeqLM,
         TFAutoModelForSequenceClassification,
-        TFAutoModelForTableQuestionAnswering,
         TFAutoModelForTokenClassification,
     )
 
@@ -93,31 +63,13 @@ if is_torch_available():
     import torch
 
     from transformers.models.auto.modeling_auto import (
-        MODEL_FOR_MASKED_LM_MAPPING,
-        MODEL_FOR_QUESTION_ANSWERING_MAPPING,
-        MODEL_FOR_SEQ_TO_SEQ_CAUSAL_LM_MAPPING,
-        MODEL_FOR_SEQUENCE_CLASSIFICATION_MAPPING,
-        MODEL_FOR_TABLE_QUESTION_ANSWERING_MAPPING,
-        MODEL_FOR_TOKEN_CLASSIFICATION_MAPPING,
         AutoModel,
-        AutoModelForAudioClassification,
         AutoModelForCausalLM,
-        AutoModelForCTC,
-        AutoModelForImageClassification,
-        AutoModelForImageSegmentation,
         AutoModelForMaskedLM,
-        AutoModelForObjectDetection,
         AutoModelForQuestionAnswering,
-        AutoModelForSemanticSegmentation,
-        AutoModelForSeq2SeqLM,
         AutoModelForSequenceClassification,
-        AutoModelForSpeechSeq2Seq,
-        AutoModelForTableQuestionAnswering,
         AutoModelForTokenClassification,
     )
-if TYPE_CHECKING:
-    from transformers.modeling_tf_utils import TFPreTrainedModel
-    from transformers.modeling_utils import PreTrainedModel
 
 from pathlib import Path
 
@@ -236,122 +188,6 @@ for task, values in SUPPORTED_TASKS.items():
         NO_TOKENIZER_TASKS.add(task)
     elif values["type"] != "multimodal":
         raise ValueError(f"SUPPORTED_TASK {task} contains invalid type {values['type']}")
-
-
-def get_supported_tasks() -> List[str]:
-    """
-    Returns a list of supported task strings.
-    """
-    supported_tasks = list(SUPPORTED_TASKS.keys()) + list(TASK_ALIASES.keys())
-    supported_tasks.sort()
-    return supported_tasks
-
-
-def get_task(model: str, use_auth_token: Optional[str] = None) -> str:
-    tmp = io.BytesIO()
-    headers = {}
-    if use_auth_token:
-        headers["Authorization"] = f"Bearer {use_auth_token}"
-
-    try:
-        http_get(f"https://huggingface.co/api/models/{model}", tmp, headers=headers)
-        tmp.seek(0)
-        body = tmp.read()
-        data = json.loads(body)
-    except Exception as e:
-        raise RuntimeError(f"Instantiating a pipeline without a task set raised an error: {e}")
-    if "pipeline_tag" not in data:
-        raise RuntimeError(
-            f"The model {model} does not seem to have a correct `pipeline_tag` set to infer the task automatically"
-        )
-    if data.get("library_name", "transformers") != "transformers":
-        raise RuntimeError(f"This model is meant to be used with {data['library_name']} not with transformers")
-    task = data["pipeline_tag"]
-    return task
-
-
-def check_task(task: str) -> Tuple[Dict, Any]:
-    """
-    Checks an incoming task string, to validate it's correct and return the default Pipeline and Model classes, and
-    default models if they exist.
-    Args:
-        task (`str`):
-            The task defining which pipeline will be returned. Currently accepted tasks are:
-            - `"audio-classification"`
-            - `"automatic-speech-recognition"`
-            - `"conversational"`
-            - `"feature-extraction"`
-            - `"fill-mask"`
-            - `"image-classification"`
-            - `"question-answering"`
-            - `"table-question-answering"`
-            - `"text2text-generation"`
-            - `"text-classification"` (alias `"sentiment-analysis"` available)
-            - `"text-generation"`
-            - `"token-classification"` (alias `"ner"` available)
-            - `"translation"`
-            - `"translation_xx_to_yy"`
-            - `"summarization"`
-            - `"zero-shot-classification"`
-            - `"zero-shot-image-classification"`
-    Returns:
-        (task_defaults`dict`, task_options: (`tuple`, None)) The actual dictionary required to initialize the pipeline
-        and some extra task options for parametrized tasks like "translation_XX_to_YY"
-    """
-    if task in TASK_ALIASES:
-        task = TASK_ALIASES[task]
-    if task in SUPPORTED_TASKS:
-        targeted_task = SUPPORTED_TASKS[task]
-        return targeted_task, None
-
-    if task.startswith("translation"):
-        tokens = task.split("_")
-        if len(tokens) == 4 and tokens[0] == "translation" and tokens[2] == "to":
-            targeted_task = SUPPORTED_TASKS["translation"]
-            return targeted_task, (tokens[1], tokens[3])
-        raise KeyError(f"Invalid translation task {task}, use 'translation_XX_to_YY' format")
-
-    raise KeyError(f"Unknown task {task}, available tasks are {get_supported_tasks() + ['translation_XX_to_YY']}")
-
-
-def _create_quantized_graph(quantizer, model, graph_path, feature):
-    logger.info(f"Creating quantized graph from {graph_path.as_posix()}")
-    quantizer.fit(model.config.name_or_path, output_dir=str(graph_path.parent.as_posix()),
-                  feature=feature)
-
-
-def _warmup_onnx_graph(self, n=10):
-    for _ in range(n):
-        self.__call__(**self.example)
-
-
-def _export_onnx_graph(quantizer, model, graph_path, feature):
-    # if graph exists, but we are here then it means something went wrong in previous load
-    # so delete old graph
-    if graph_path.exists():
-        graph_path.unlink()
-
-    # create parent dir
-    if not graph_path.parent.exists():
-        os.makedirs(graph_path.parent.as_posix())
-
-    logger.info(f"Saving onnx graph at {graph_path.as_posix()}")
-
-    quantizer.export(model.config.name_or_path, output_path=graph_path,
-                     feature=feature)
-
-def create_model_for_providers(model_path: str) -> InferenceSession:
-    logger.info(f"Creating model for providers: {model_path}")
-    # Few properties that might have an impact on performances (provided by MS)
-    options = SessionOptions()
-    options.intra_op_num_threads = 1
-    options.graph_optimization_level = GraphOptimizationLevel.ORT_ENABLE_ALL
-
-    # Load the model as a graph and prepare the CPU backend
-    session = InferenceSession(str(model_path), options)
-    session.disable_fallback()
-
-    return session
 
 
 def pipeline(
@@ -684,11 +520,6 @@ def pipeline(
         model=model,
         framework=framework,
         task=task,
-        # use_onnx=use_onnx,
-        # graph_path=graph_path,
-        # optimize=optimize,
-        # ort_config=ort_config,
-        # feature=targeted_task["feature"],
         example=targeted_task["example"],
         onnx_model=onnx_model,
         **kwargs
