+++ optimum_transformers/pipelines/base.py
@@ -1,680 +1,41 @@
-# coding=utf-8
-# Copyright 2018 The HuggingFace Inc. team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-import collections
-import csv
-import importlib
-import json
+from onnxruntime import GraphOptimizationLevel, InferenceSession, SessionOptions
+from transformers.utils import logging
 import os
-import pickle
-import sys
-import types
-import warnings
-from abc import ABC, abstractmethod
-from collections import UserDict
-from contextlib import contextmanager
-from os.path import abspath, exists
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Union
-from pathlib import Path
-
-from packaging import version
-
-from transformers.feature_extraction_utils import PreTrainedFeatureExtractor
-from transformers.file_utils import ModelOutput, add_end_docstrings, is_tf_available, is_torch_available
-from transformers.modelcard import ModelCard
-from transformers.models.auto.configuration_auto import AutoConfig
-from transformers.tokenization_utils import PreTrainedTokenizer
-from transformers.utils import logging
-
-import numpy as np
-from onnxruntime import GraphOptimizationLevel, InferenceSession, SessionOptions, get_all_providers
-from onnxruntime.transformers import optimizer
-from onnxruntime.transformers.onnx_model_bert import BertOptimizationOptions, BertOnnxModel
-from onnxruntime.transformers.fusion_options import FusionOptions
-
-# FusionOptions
-from psutil import cpu_count
-from transformers import AutoConfig, SquadFeatures
-from transformers.configuration_utils import PretrainedConfig
-from transformers.convert_graph_to_onnx import convert_pytorch, convert_tensorflow, infer_shapes, convert
-from transformers.data import SquadExample, squad_convert_examples_to_features
-from transformers.file_utils import add_end_docstrings, is_tf_available, is_torch_available
-from transformers.modelcard import ModelCard
-from transformers import AutoTokenizer
-from transformers import BasicTokenizer
-from transformers.tokenization_utils import PreTrainedTokenizer
-from transformers.tokenization_utils_base import BatchEncoding, PaddingStrategy
-from transformers.utils import logging
-from onnxruntime.quantization import quantize_dynamic, QuantType, quantize_static, quantize
-from optimum.onnxruntime import ORTConfig, ORTQuantizer
-
-GenericTensor = Union[List["GenericTensor"], "torch.Tensor", "tf.Tensor"]
-
-if is_tf_available():
-    import tensorflow as tf
-
-    from transformers.models.auto.modeling_tf_auto import TFAutoModel
-
-if is_torch_available():
-    import torch
-    from torch.utils.data import DataLoader, Dataset
-
-    from transformers.models.auto.modeling_auto import AutoModel
-
-    # Re-export for backward compatibility
-    from transformers.pipelines.pt_utils import KeyDataset
-else:
-    Dataset = None
-    KeyDataset = None
-
-if TYPE_CHECKING:
-    from transformers.modeling_tf_utils import TFPreTrainedModel
-    from transformers.modeling_utils import PreTrainedModel
 
 logger = logging.get_logger(__name__)
 
 
-def no_collate_fn(items):
-    if len(items) != 1:
-        raise ValueError("This collate_fn is meant to be used with batch_size=1")
-    return items[0]
+def _create_quantized_graph(quantizer, model, graph_path, feature):
+    logger.info(f"Creating quantized graph from {graph_path.as_posix()}")
+    quantizer.fit(model.config.name_or_path, output_dir=str(graph_path.parent.as_posix()),
+                  feature=feature)
 
 
-def _pad(items, key, padding_value, padding_side):
-    batch_size = len(items)
-    if isinstance(items[0][key], torch.Tensor):
-        # Others include `attention_mask` etc...
-        shape = items[0][key].shape
-        dim = len(shape)
-        if dim == 4:
-            # This is probable image so padding shouldn't be necessary
-            # B, C, H, W
-            return torch.cat([item[key] for item in items], dim=0)
-        max_length = max(item[key].shape[1] for item in items)
-        dtype = items[0][key].dtype
+def _warmup_onnx_graph(self, n=10):
+    for _ in range(n):
+        self.__call__(**self.example)
 
-        if dim == 2:
-            tensor = torch.zeros((batch_size, max_length), dtype=dtype) + padding_value
-        elif dim == 3:
-            tensor = torch.zeros((batch_size, max_length, shape[-1]), dtype=dtype) + padding_value
 
-        for i, item in enumerate(items):
-            if dim == 2:
-                if padding_side == "left":
-                    tensor[i, -len(item[key][0]):] = item[key][0].clone()
-                else:
-                    tensor[i, : len(item[key][0])] = item[key][0].clone()
-            elif dim == 3:
-                if padding_side == "left":
-                    tensor[i, -len(item[key][0]):, :] = item[key][0].clone()
-                else:
-                    tensor[i, : len(item[key][0]), :] = item[key][0].clone()
-        return tensor
-    else:
-        return [item[key] for item in items]
+def _forward_onnx(onnx_model, inputs, return_tensors=False):
+    inputs_onnx = {k: v.cpu().detach().numpy() for k, v in inputs.items()}
+    predictions = onnx_model.run(None, inputs_onnx)
+    return predictions
 
 
-def pad_collate_fn(tokenizer, feature_extractor):
-    # Tokenizer
-    t_padding_side = None
-    # Feature extractor
-    f_padding_side = None
-    if tokenizer is None and feature_extractor is None:
-        raise ValueError("Pipeline without tokenizer or feature_extractor cannot do batching")
-    if tokenizer is not None:
-        if tokenizer.pad_token_id is None:
-            raise ValueError(
-                "Pipeline with tokenizer without pad_token cannot do batching. You can try to set it with "
-                "`pipe.tokenizer.pad_token_id = model.config.eos_token_id`."
-            )
-        else:
-            t_padding_value = tokenizer.pad_token_id
-            t_padding_side = tokenizer.padding_side
-    if feature_extractor is not None:
-        # Feature extractor can be images, where no padding is expected
-        f_padding_value = getattr(feature_extractor, "padding_value", None)
-        f_padding_side = getattr(feature_extractor, "padding_side", None)
+def _export_onnx_graph(quantizer, model, graph_path, feature):
+    # if graph exists, but we are here then it means something went wrong in previous load
+    # so delete old graph
+    if graph_path.exists():
+        graph_path.unlink()
 
-    if t_padding_side is not None and f_padding_side is not None and t_padding_side != f_padding_side:
-        raise ValueError(
-            f"The feature extractor, and tokenizer don't agree on padding side {t_padding_side} != {f_padding_side}"
-        )
-    padding_side = "right"
-    if t_padding_side is not None:
-        padding_side = t_padding_side
-    if f_padding_side is not None:
-        padding_side = f_padding_side
+    # create parent dir
+    if not graph_path.parent.exists():
+        os.makedirs(graph_path.parent.as_posix())
 
-    def inner(items):
-        keys = set(items[0].keys())
-        for item in items:
-            if set(item.keys()) != keys:
-                raise ValueError(
-                    f"The elements of the batch contain different keys. Cannot batch them ({set(item.keys())} != {keys})"
-                )
-        # input_values, input_pixels, input_ids, ...
-        padded = {}
-        for key in keys:
-            if key in {"input_ids"}:
-                _padding_value = t_padding_value
-            elif key in {"input_values", "pixel_values", "input_features"}:
-                _padding_value = f_padding_value
-            elif key in {"p_mask", "special_tokens_mask"}:
-                _padding_value = 1
-            elif key in {"attention_mask", "token_type_ids"}:
-                _padding_value = 0
-            else:
-                # This is likely another random key maybe even user provided
-                _padding_value = 0
-            padded[key] = _pad(items, key, _padding_value, padding_side)
-        return padded
+    logger.info(f"Saving onnx graph at {graph_path.as_posix()}")
 
-    return inner
-
-
-def infer_framework_load_model(
-        model,
-        config: AutoConfig,
-        model_classes: Optional[Dict[str, Tuple[type]]] = None,
-        task: Optional[str] = None,
-        framework: Optional[str] = None,
-        **model_kwargs
-):
-    """
-    Select framework (TensorFlow or PyTorch) to use from the `model` passed. Returns a tuple (framework, model).
-    If `model` is instantiated, this function will just infer the framework from the model class. Otherwise `model` is
-    actually a checkpoint name and this method will try to instantiate it using `model_classes`. Since we don't want to
-    instantiate the model twice, this model is returned for use by the pipeline.
-    If both frameworks are installed and available for `model`, PyTorch is selected.
-    Args:
-        model (`str`, [`PreTrainedModel`] or [`TFPreTrainedModel`]):
-            The model to infer the framework from. If `str`, a checkpoint name. The model to infer the framewrok from.
-        config ([`AutoConfig`]):
-            The config associated with the model to help using the correct class
-        model_classes (dictionary `str` to `type`, *optional*):
-            A mapping framework to class.
-        task (`str`):
-            The task defining which pipeline will be returned.
-        model_kwargs:
-            Additional dictionary of keyword arguments passed along to the model's `from_pretrained(...,
-            **model_kwargs)` function.
-    Returns:
-        `Tuple`: A tuple framework, model.
-    """
-    if not is_tf_available() and not is_torch_available():
-        raise RuntimeError(
-            "At least one of TensorFlow 2.0 or PyTorch should be installed. "
-            "To install TensorFlow 2.0, read the instructions at https://www.tensorflow.org/install/ "
-            "To install PyTorch, read the instructions at https://pytorch.org/."
-        )
-    if isinstance(model, str):
-        model_kwargs["_from_pipeline"] = task
-        class_tuple = ()
-        look_pt = is_torch_available() and framework in {"pt", None}
-        look_tf = is_tf_available() and framework in {"tf", None}
-        if model_classes:
-            if look_pt:
-                class_tuple = class_tuple + model_classes.get("pt", (AutoModel,))
-            if look_tf:
-                class_tuple = class_tuple + model_classes.get("tf", (TFAutoModel,))
-        if config.architectures:
-            classes = []
-            for architecture in config.architectures:
-                transformers_module = importlib.import_module("transformers")
-                if look_pt:
-                    _class = getattr(transformers_module, architecture, None)
-                    if _class is not None:
-                        classes.append(_class)
-                if look_tf:
-                    _class = getattr(transformers_module, f"TF{architecture}", None)
-                    if _class is not None:
-                        classes.append(_class)
-            class_tuple = class_tuple + tuple(classes)
-
-        if len(class_tuple) == 0:
-            raise ValueError(f"Pipeline cannot infer suitable model classes from {model}")
-
-        for model_class in class_tuple:
-            kwargs = model_kwargs.copy()
-            if framework == "pt" and model.endswith(".h5"):
-                kwargs["from_tf"] = True
-                logger.warning(
-                    "Model might be a TensorFlow model (ending with `.h5`) but TensorFlow is not available. "
-                    "Trying to load the model with PyTorch."
-                )
-            elif framework == "tf" and model.endswith(".bin"):
-                kwargs["from_pt"] = True
-                logger.warning(
-                    "Model might be a PyTorch model (ending with `.bin`) but PyTorch is not available. "
-                    "Trying to load the model with Tensorflow."
-                )
-
-            try:
-                model = model_class.from_pretrained(model, **kwargs)
-                if hasattr(model, "eval"):
-                    model = model.eval()
-                # Stop loading on the first successful load.
-                break
-            except (OSError, ValueError):
-                continue
-
-        if isinstance(model, str):
-            raise ValueError(f"Could not load model {model} with any of the following classes: {class_tuple}.")
-
-    framework = "tf" if model.__class__.__name__.startswith("TF") else "pt"
-    return framework, model
-
-
-def infer_framework_from_model(
-        model,
-        model_classes: Optional[Dict[str, Tuple[type]]] = None,
-        task: Optional[str] = None,
-        framework: Optional[str] = None,
-        **model_kwargs
-):
-    """
-    Select framework (TensorFlow or PyTorch) to use from the `model` passed. Returns a tuple (framework, model).
-    If `model` is instantiated, this function will just infer the framework from the model class. Otherwise `model` is
-    actually a checkpoint name and this method will try to instantiate it using `model_classes`. Since we don't want to
-    instantiate the model twice, this model is returned for use by the pipeline.
-    If both frameworks are installed and available for `model`, PyTorch is selected.
-    Args:
-        model (`str`, [`PreTrainedModel`] or [`TFPreTrainedModel`]):
-            The model to infer the framework from. If `str`, a checkpoint name. The model to infer the framewrok from.
-        model_classes (dictionary `str` to `type`, *optional*):
-            A mapping framework to class.
-        task (`str`):
-            The task defining which pipeline will be returned.
-        model_kwargs:
-            Additional dictionary of keyword arguments passed along to the model's `from_pretrained(...,
-            **model_kwargs)` function.
-    Returns:
-        `Tuple`: A tuple framework, model.
-    """
-    if isinstance(model, str):
-        config = AutoConfig.from_pretrained(model, _from_pipeline=task, **model_kwargs)
-    else:
-        config = model.config
-    return infer_framework_load_model(
-        model, config, model_classes=model_classes, _from_pipeline=task, task=task, framework=framework, **model_kwargs
-    )
-
-
-def get_framework(model, revision: Optional[str] = None):
-    """
-    Select framework (TensorFlow or PyTorch) to use.
-    Args:
-        model (`str`, [`PreTrainedModel`] or [`TFPreTrainedModel`]):
-            If both frameworks are installed, picks the one corresponding to the model passed (either a model class or
-            the model name). If no specific model is provided, defaults to using PyTorch.
-    """
-    warnings.warn(
-        "`get_framework` is deprecated and will be removed in v5, use `infer_framework_from_model` instead.",
-        FutureWarning,
-    )
-    if not is_tf_available() and not is_torch_available():
-        raise RuntimeError(
-            "At least one of TensorFlow 2.0 or PyTorch should be installed. "
-            "To install TensorFlow 2.0, read the instructions at https://www.tensorflow.org/install/ "
-            "To install PyTorch, read the instructions at https://pytorch.org/."
-        )
-    if isinstance(model, str):
-        if is_torch_available() and not is_tf_available():
-            model = AutoModel.from_pretrained(model, revision=revision)
-        elif is_tf_available() and not is_torch_available():
-            model = TFAutoModel.from_pretrained(model, revision=revision)
-        else:
-            try:
-                model = AutoModel.from_pretrained(model, revision=revision)
-            except OSError:
-                model = TFAutoModel.from_pretrained(model, revision=revision)
-
-    framework = "tf" if model.__class__.__name__.startswith("TF") else "pt"
-    return framework
-
-
-def get_default_model(targeted_task: Dict, framework: Optional[str], task_options: Optional[Any]) -> str:
-    """
-    Select a default model to use for a given task. Defaults to pytorch if ambiguous.
-    Args:
-        targeted_task (`Dict` ):
-           Dictionary representing the given task, that should contain default models
-        framework (`str`, None)
-           "pt", "tf" or None, representing a specific framework if it was specified, or None if we don't know yet.
-        task_options (`Any`, None)
-           Any further value required by the task to get fully specified, for instance (SRC, TGT) languages for
-           translation task.
-    Returns
-        `str` The model string representing the default model for this pipeline
-    """
-    if is_torch_available() and not is_tf_available():
-        framework = "pt"
-    elif is_tf_available() and not is_torch_available():
-        framework = "tf"
-
-    defaults = targeted_task["default"]
-    if task_options:
-        if task_options not in defaults:
-            raise ValueError(f"The task does not provide any default models for options {task_options}")
-        default_models = defaults[task_options]["model"]
-    elif "model" in defaults:
-        default_models = targeted_task["default"]["model"]
-    else:
-        # XXX This error message needs to be updated to be more generic if more tasks are going to become
-        # parametrized
-        raise ValueError('The task defaults can\'t be correctly selected. You probably meant "translation_XX_to_YY"')
-
-    if framework is None:
-        framework = "pt"
-
-    return default_models[framework]
-
-
-class PipelineException(Exception):
-    """
-    Raised by a [`Pipeline`] when handling __call__.
-    Args:
-        task (`str`): The task of the pipeline.
-        model (`str`): The model used by the pipeline.
-        reason (`str`): The error message to display.
-    """
-
-    def __init__(self, task: str, model: str, reason: str):
-        super().__init__(reason)
-
-        self.task = task
-        self.model = model
-
-
-class ArgumentHandler(ABC):
-    """
-    Base interface for handling arguments for each [`~pipelines.Pipeline`].
-    """
-
-    @abstractmethod
-    def __call__(self, *args, **kwargs):
-        raise NotImplementedError()
-
-
-class PipelineDataFormat:
-    """
-    Base class for all the pipeline supported data format both for reading and writing. Supported data formats
-    currently includes:
-    - JSON
-    - CSV
-    - stdin/stdout (pipe)
-    `PipelineDataFormat` also includes some utilities to work with multi-columns like mapping from datasets columns to
-    pipelines keyword arguments through the `dataset_kwarg_1=dataset_column_1` format.
-    Args:
-        output_path (`str`, *optional*): Where to save the outgoing data.
-        input_path (`str`, *optional*): Where to look for the input data.
-        column (`str`, *optional*): The column to read.
-        overwrite (`bool`, *optional*, defaults to `False`):
-            Whether or not to overwrite the `output_path`.
-    """
-
-    SUPPORTED_FORMATS = ["json", "csv", "pipe"]
-
-    def __init__(
-            self,
-            output_path: Optional[str],
-            input_path: Optional[str],
-            column: Optional[str],
-            overwrite: bool = False,
-    ):
-        self.output_path = output_path
-        self.input_path = input_path
-        self.column = column.split(",") if column is not None else [""]
-        self.is_multi_columns = len(self.column) > 1
-
-        if self.is_multi_columns:
-            self.column = [tuple(c.split("=")) if "=" in c else (c, c) for c in self.column]
-
-        if output_path is not None and not overwrite:
-            if exists(abspath(self.output_path)):
-                raise OSError(f"{self.output_path} already exists on disk")
-
-        if input_path is not None:
-            if not exists(abspath(self.input_path)):
-                raise OSError(f"{self.input_path} doesnt exist on disk")
-
-    @abstractmethod
-    def __iter__(self):
-        raise NotImplementedError()
-
-    @abstractmethod
-    def save(self, data: Union[dict, List[dict]]):
-        """
-        Save the provided data object with the representation for the current [`~pipelines.PipelineDataFormat`].
-        Args:
-            data (`dict` or list of `dict`): The data to store.
-        """
-        raise NotImplementedError()
-
-    def save_binary(self, data: Union[dict, List[dict]]) -> str:
-        """
-        Save the provided data object as a pickle-formatted binary data on the disk.
-        Args:
-            data (`dict` or list of `dict`): The data to store.
-        Returns:
-            `str`: Path where the data has been saved.
-        """
-        path, _ = os.path.splitext(self.output_path)
-        binary_path = os.path.extsep.join((path, "pickle"))
-
-        with open(binary_path, "wb+") as f_output:
-            pickle.dump(data, f_output)
-
-        return binary_path
-
-    @staticmethod
-    def from_str(
-            format: str,
-            output_path: Optional[str],
-            input_path: Optional[str],
-            column: Optional[str],
-            overwrite=False,
-    ) -> "PipelineDataFormat":
-        """
-        Creates an instance of the right subclass of [`~pipelines.PipelineDataFormat`] depending on `format`.
-        Args:
-            format: (`str`):
-                The format of the desired pipeline. Acceptable values are `"json"`, `"csv"` or `"pipe"`.
-            output_path (`str`, *optional*):
-                Where to save the outgoing data.
-            input_path (`str`, *optional*):
-                Where to look for the input data.
-            column (`str`, *optional*):
-                The column to read.
-            overwrite (`bool`, *optional*, defaults to `False`):
-                Whether or not to overwrite the `output_path`.
-        Returns:
-            [`~pipelines.PipelineDataFormat`]: The proper data format.
-        """
-        if format == "json":
-            return JsonPipelineDataFormat(output_path, input_path, column, overwrite=overwrite)
-        elif format == "csv":
-            return CsvPipelineDataFormat(output_path, input_path, column, overwrite=overwrite)
-        elif format == "pipe":
-            return PipedPipelineDataFormat(output_path, input_path, column, overwrite=overwrite)
-        else:
-            raise KeyError(f"Unknown reader {format} (Available reader are json/csv/pipe)")
-
-
-class CsvPipelineDataFormat(PipelineDataFormat):
-    """
-    Support for pipelines using CSV data format.
-    Args:
-        output_path (`str`, *optional*): Where to save the outgoing data.
-        input_path (`str`, *optional*): Where to look for the input data.
-        column (`str`, *optional*): The column to read.
-        overwrite (`bool`, *optional*, defaults to `False`):
-            Whether or not to overwrite the `output_path`.
-    """
-
-    def __init__(
-            self,
-            output_path: Optional[str],
-            input_path: Optional[str],
-            column: Optional[str],
-            overwrite=False,
-    ):
-        super().__init__(output_path, input_path, column, overwrite=overwrite)
-
-    def __iter__(self):
-        with open(self.input_path, "r") as f:
-            reader = csv.DictReader(f)
-            for row in reader:
-                if self.is_multi_columns:
-                    yield {k: row[c] for k, c in self.column}
-                else:
-                    yield row[self.column[0]]
-
-    def save(self, data: List[dict]):
-        """
-        Save the provided data object with the representation for the current [`~pipelines.PipelineDataFormat`].
-        Args:
-            data (`List[dict]`): The data to store.
-        """
-        with open(self.output_path, "w") as f:
-            if len(data) > 0:
-                writer = csv.DictWriter(f, list(data[0].keys()))
-                writer.writeheader()
-                writer.writerows(data)
-
-
-class JsonPipelineDataFormat(PipelineDataFormat):
-    """
-    Support for pipelines using JSON file format.
-    Args:
-        output_path (`str`, *optional*): Where to save the outgoing data.
-        input_path (`str`, *optional*): Where to look for the input data.
-        column (`str`, *optional*): The column to read.
-        overwrite (`bool`, *optional*, defaults to `False`):
-            Whether or not to overwrite the `output_path`.
-    """
-
-    def __init__(
-            self,
-            output_path: Optional[str],
-            input_path: Optional[str],
-            column: Optional[str],
-            overwrite=False,
-    ):
-        super().__init__(output_path, input_path, column, overwrite=overwrite)
-
-        with open(input_path, "r") as f:
-            self._entries = json.load(f)
-
-    def __iter__(self):
-        for entry in self._entries:
-            if self.is_multi_columns:
-                yield {k: entry[c] for k, c in self.column}
-            else:
-                yield entry[self.column[0]]
-
-    def save(self, data: dict):
-        """
-        Save the provided data object in a json file.
-        Args:
-            data (`dict`): The data to store.
-        """
-        with open(self.output_path, "w") as f:
-            json.dump(data, f)
-
-
-class PipedPipelineDataFormat(PipelineDataFormat):
-    """
-    Read data from piped input to the python process. For multi columns data, columns should separated by \t
-    If columns are provided, then the output will be a dictionary with {column_x: value_x}
-    Args:
-        output_path (`str`, *optional*): Where to save the outgoing data.
-        input_path (`str`, *optional*): Where to look for the input data.
-        column (`str`, *optional*): The column to read.
-        overwrite (`bool`, *optional*, defaults to `False`):
-            Whether or not to overwrite the `output_path`.
-    """
-
-    def __iter__(self):
-        for line in sys.stdin:
-            # Split for multi-columns
-            if "\t" in line:
-
-                line = line.split("\t")
-                if self.column:
-                    # Dictionary to map arguments
-                    yield {kwargs: l for (kwargs, _), l in zip(self.column, line)}
-                else:
-                    yield tuple(line)
-
-            # No dictionary to map arguments
-            else:
-                yield line
-
-    def save(self, data: dict):
-        """
-        Print the data.
-        Args:
-            data (`dict`): The data to store.
-        """
-        print(data)
-
-    def save_binary(self, data: Union[dict, List[dict]]) -> str:
-        if self.output_path is None:
-            raise KeyError(
-                "When using piped input on pipeline outputting large object requires an output file path. "
-                "Please provide such output path through --output argument."
-            )
-
-        return super().save_binary(data)
-
-
-
-
-PIPELINE_INIT_ARGS = r"""
-    Arguments:
-        model ([`PreTrainedModel`] or [`TFPreTrainedModel`]):
-            The model that will be used by the pipeline to make predictions. This needs to be a model inheriting from
-            [`PreTrainedModel`] for PyTorch and [`TFPreTrainedModel`] for TensorFlow.
-        tokenizer ([`PreTrainedTokenizer`]):
-            The tokenizer that will be used by the pipeline to encode data for the model. This object inherits from
-            [`PreTrainedTokenizer`].
-        modelcard (`str` or [`ModelCard`], *optional*):
-            Model card attributed to the model for this pipeline.
-        framework (`str`, *optional*):
-            The framework to use, either `"pt"` for PyTorch or `"tf"` for TensorFlow. The specified framework must be
-            installed.
-            If no framework is specified, will default to the one currently installed. If no framework is specified and
-            both frameworks are installed, will default to the framework of the `model`, or to PyTorch if no model is
-            provided.
-        task (`str`, defaults to `""`):
-            A task-identifier for the pipeline.
-        num_workers (`int`, *optional*, defaults to 8):
-            When the pipeline will use *DataLoader* (when passing a dataset, on GPU for a Pytorch model), the number of
-            workers to be used.
-        batch_size (`int`, *optional*, defaults to 1):
-            When the pipeline will use *DataLoader* (when passing a dataset, on GPU for a Pytorch model), the size of
-            the batch to use, for inference this is not always beneficial, please read [Batching with
-            pipelines](https://huggingface.co/transformers/main_classes/pipelines.html#pipeline-batching) .
-        args_parser ([`~pipelines.ArgumentHandler`], *optional*):
-            Reference to the object in charge of parsing supplied pipeline parameters.
-        device (`int`, *optional*, defaults to -1):
-            Device ordinal for CPU/GPU supports. Setting this to -1 will leverage CPU, a positive will run the model on
-            the associated CUDA device id.
-        binary_output (`bool`, *optional*, defaults to `False`):
-            Flag indicating if the output the pipeline should happen in a binary format (i.e., pickle) or as raw text.
-"""
+    quantizer.export(model.config.name_or_path, output_path=graph_path,
+                     feature=feature)
 
 
 def create_model_for_providers(model_path: str) -> InferenceSession:
@@ -689,14 +50,3 @@ def create_model_for_providers(model_path: str) -> InferenceSession:
     session.disable_fallback()
 
     return session
-
-
-def _forward_onnx(onnx_model, inputs, return_tensors=False):
-    inputs_onnx = {k: v.cpu().detach().numpy() for k, v in inputs.items()}
-    predictions = onnx_model.run(None, inputs_onnx)
-    return predictions
-
-
-def _warmup_onnx_graph(pipe, n=10):
-    for _ in range(n):
-        pipe.__call__(*pipe.example)
