+++ optimum_transformers/pipelines/feature_extraction.py
@@ -1,5 +1,5 @@
 from transformers.modeling_outputs import BaseModelOutput
-from .base import _forward_onnx, _warmup_onnx_graph
+from . import _forward_onnx, _warmup_onnx_graph
 from transformers import FeatureExtractionPipeline
 import torch
